+++ Src/Void.Engine/Assets/Loaders/Sound.cs
@@ -93,8 +93,11 @@ public enum SoundPriority
 /// </para>
 /// <para>
 /// <b>Asset Caching and Eviction:</b>
-/// Sound assets are cached by the AssetManager and automatically evicted after
-/// a configurable idle time (<see cref="GameSettings.AssetEvictionMinutes"/>).
+/// Sound assets are cached by the AssetManager and automatically evicted when
+/// they haven't been accessed for a configurable number of asset accesses
+/// (<see cref="GameSettings.AssetStalenessThreshold"/>). The eviction system
+/// uses a global access counter to track when assets were last used. When
+/// <see cref="CreateInstance"/> is called, the sound's access tick is updated.
 /// When evicted, <see cref="Unload()"/> is called to release the audio buffer.
 /// The asset will be reloaded automatically if <see cref="CreateInstance"/>
 /// is called while unloaded.
@@ -142,7 +145,7 @@ public sealed class Sound : IAsset
     /// <summary>
     /// Gets the last access time of the asset for eviction tracking.
     /// </summary>
-    public ushort LastAccessTick { get; set; }
+    public uint LastAccessTick { get; set; }
 
     /// <summary>
     /// Gets the underlying SFML sound buffer containing the decoded audio data.
@@ -272,7 +275,7 @@ public sealed class Sound : IAsset
                 Load();
             }
 
-           AssetManager.Instance.Touch(this);
+            AssetManager.Instance.Touch(this);
 
             var instance = SoundInstancePool.Instance.GetInstance();
